file_decryptor into a seperate module
