+++ parquet/src/encryption/mod.rs
@@ -19,3 +19,4 @@
 //! in the [spec](https://github.com/apache/parquet-format/blob/master/Encryption.md).
 
 pub mod ciphers;
+pub mod decryption;
